requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
asttokens=2.0.5=pypi_0
attrs=21.4.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.0=pypi_0
brotlipy=0.7.0=py39h2bbff1b_1003
ca-certificates=2022.3.29=haa95532_1
certifi=2021.10.8=py39haa95532_2
cffi=1.15.0=py39h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.4=pypi_0
cryptography=36.0.0=py39h21b164f_0
cuda=11.3.1=h46d18f5_0
cuda-command-line-tools=11.3.1=h46d18f5_0
cuda-compiler=11.3.1=h46d18f5_0
cuda-cudart=11.3.109=h90d3b27_0
cuda-cuobjdump=11.3.122=hbfb3ce2_0
cuda-cupti=11.3.111=h1d2d2a3_0
cuda-cuxxfilt=11.3.122=hc6e7a92_0
cuda-libraries=11.3.1=h46d18f5_0
cuda-libraries-dev=11.3.1=h46d18f5_0
cuda-memcheck=11.3.109=h23f69e8_0
cuda-nvcc=11.3.122=h790fb98_0
cuda-nvdisasm=11.3.122=h2cc5393_0
cuda-nvml-dev=11.3.58=hbc9c638_0
cuda-nvprof=11.3.111=h418b846_0
cuda-nvprune=11.3.122=h8d04b39_0
cuda-nvrtc=11.3.122=h4fd7f55_0
cuda-nvtx=11.3.109=hc90bad4_0
cuda-runtime=11.3.1=h46d18f5_0
cuda-samples=11.3.58=h9a5194a_0
cuda-sanitizer-api=11.3.111=hb9b8c8e_0
cuda-thrust=11.3.109=hfc8979d_0
cuda-toolkit=11.3.1=h46d18f5_0
cuda-tools=11.3.1=h46d18f5_0
cudatoolkit=11.3.1=h59b6b97_2
cudnn=8.2.1=cuda11.3_0
cycler=0.11.0=pypi_0
debugpy=1.6.0=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
entrypoints=0.4=pypi_0
executing=0.8.3=pypi_0
fastjsonschema=2.15.3=pypi_0
fonttools=4.33.3=pypi_0
freetype=2.10.4=hd328e21_0
idna=3.3=pyhd3eb1b0_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.13.0=pypi_0
ipython=8.3.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.0=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h2bbff1b_0
jsonschema=4.4.0=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=7.3.0=pypi_0
jupyter-console=6.4.3=pypi_0
jupyter-core=4.10.0=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.0=pypi_0
kiwisolver=1.4.2=pypi_0
libcublas=11.5.1.109=h534c3b9_0
libcufft=10.4.2.109=h5b78ae5_0
libcurand=10.2.4.109=h8bccdc1_0
libcusolver=11.1.2.109=hbc28524_0
libcusparse=11.6.0.109=h4f1f2d5_0
libnpp=11.3.3.95=h0a60a38_0
libnvjpeg=11.5.0.109=ha56d775_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.2.0=hd0e1b90_0
libuv=1.40.0=he774522_0
libwebp=1.2.2=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
markupsafe=2.1.1=pypi_0
matplotlib=3.5.1=pypi_0
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py39h2bbff1b_0
mkl_fft=1.3.1=py39h277e83a_0
mkl_random=1.2.2=py39hf11a4ad_0
natsort=8.2.0=pypi_0
nbclient=0.6.0=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.3.0=pypi_0
nest-asyncio=1.5.5=pypi_0
notebook=6.4.11=pypi_0
numpy=1.21.5=py39h7a0a035_1
numpy-base=1.21.5=py39hca35cd5_1
opencv-python=4.6.0.66=pypi_0
openssl=1.1.1n=h2bbff1b_0
packaging=21.3=pypi_0
pandas=1.4.2=pypi_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pickleshare=0.7.5=pypi_0
pillow=9.0.1=py39hdc2b20a_0
pip=21.2.4=py39haa95532_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.29=pypi_0
psutil=5.9.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.12.0=pypi_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.8=pypi_0
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.12=h6244533_0
python-dateutil=2.8.2=pypi_0
pytorch=1.11.0=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=pypi_0
pywin32=304=pypi_0
pywinpty=2.0.5=pypi_0
pyzmq=22.3.0=pypi_0
qtconsole=5.3.0=pypi_0
qtpy=2.0.1=pypi_0
requests=2.27.1=pyhd3eb1b0_0
scikit-learn=1.1.2=pypi_0
scipy=1.9.1=pypi_0
seaborn=0.12.0=pypi_0
send2trash=1.8.0=pypi_0
setuptools=61.2.0=py39haa95532_0
six=1.16.0=pyhd3eb1b0_1
sklearn=0.0=pypi_0
soupsieve=2.3.2.post1=pypi_0
sqlite=3.38.2=h2bbff1b_0
stack-data=0.2.0=pypi_0
terminado=0.13.3=pypi_0
threadpoolctl=3.1.0=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h2bbff1b_0
torchaudio=0.11.0=py39_cu113
torchvision=0.12.0=py39_cu113
tornado=6.1=pypi_0
tqdm=4.64.1=pypi_0
traitlets=5.1.1=pypi_0
typing_extensions=4.1.1=pyh06a4308_0
tzdata=2022a=hda174b7_0
urllib3=1.26.9=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.0=pypi_0
win_inet_pton=1.1.0=py39haa95532_0
wincertstore=0.2=py39haa95532_2
xz=5.2.5=h62dcd97_0
zlib=1.2.12=h8cc25b3_2
zstd=1.4.9=h19a0ad4_0
